Move new ES client and reader to different package
