+++ cmd/opentelemetry/app/internal/esclient/client_test.go
