+++ cmd/opentelemetry/app/internal/esclient/es6client.go
