+++ cmd/opentelemetry/app/internal/esclient/es6client_test.go
