+++ cmd/opentelemetry/app/internal/esclient/es7client.go
