+++ cmd/opentelemetry/app/internal/esclient/ping.go
