+++ cmd/opentelemetry/app/internal/esclient/ping_test.go
