+++ cmd/opentelemetry/app/reader/es/esdependencyreader/dependency_store.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package dependencystore
+package esdependencyreader
 
 import (
 	"bytes"
@@ -23,7 +23,7 @@ import (
 
 	"go.uber.org/zap"
 
-	"github.com/jaegertracing/jaeger/cmd/opentelemetry/app/exporter/elasticsearchexporter/esclient"
+	"github.com/jaegertracing/jaeger/cmd/opentelemetry/app/internal/esclient"
 	"github.com/jaegertracing/jaeger/model"
 	"github.com/jaegertracing/jaeger/plugin/storage/es/dependencystore/dbmodel"
 	"github.com/jaegertracing/jaeger/storage/dependencystore"
