+++ cmd/opentelemetry/app/reader/es/esspanreader/query_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package spanreader
+package esspanreader
 
 import (
 	"encoding/json"
